requirements.txt
numpy
pandas
scikit-learn
matplotlib
pyserial
tensorflow
joblib
streamlit
